+++ packages/eslint-plugin/src/rules/no-array-delete.ts
@@ -5,7 +5,7 @@ import { AST_NODE_TYPES, AST_TOKEN_TYPES } from '@typescript-eslint/utils';
 
 import {
   createRule,
-  DEPRECATED_getConstrainedTypeAtLocation,
+  getConstrainedTypeAtLocation,
   getParserServices,
 } from '../util';
 
@@ -58,10 +58,7 @@ export default createRule<[], MessageId>({
           return;
         }
 
-        const type = DEPRECATED_getConstrainedTypeAtLocation(
-          services,
-          argument.object,
-        );
+        const type = getConstrainedTypeAtLocation(services, argument.object);
 
         if (!isUnderlyingTypeArray(type)) {
           return;
