+++ packages/eslint-plugin/src/rules/no-base-to-string.ts
@@ -6,7 +6,7 @@ import * as ts from 'typescript';
 
 import {
   createRule,
-  DEPRECATED_getConstrainedTypeAtLocation,
+  getConstrainedTypeAtLocation,
   getParserServices,
   getTypeName,
   nullThrows,
@@ -274,10 +274,7 @@ export default createRule<Options, MessageIds>({
         node: TSESTree.Expression,
       ): void {
         const memberExpr = node.parent as TSESTree.MemberExpression;
-        const type = DEPRECATED_getConstrainedTypeAtLocation(
-          services,
-          memberExpr.object,
-        );
+        const type = getConstrainedTypeAtLocation(services, memberExpr.object);
         checkExpressionForArrayJoin(memberExpr.object, type);
       },
       'CallExpression > MemberExpression.callee > Identifier[name = /^(toLocaleString|toString)$/].property'(
