+++ packages/eslint-plugin/src/rules/no-unnecessary-template-expression.ts
@@ -5,7 +5,7 @@ import * as ts from 'typescript';
 
 import {
   createRule,
-  DEPRECATED_getConstrainedTypeAtLocation,
+  getConstrainedTypeAtLocation,
   getMovedNodeCode,
   getParserServices,
   isTypeFlagSet,
@@ -51,10 +51,7 @@ export default createRule<[], MessageId>({
     function isUnderlyingTypeString(
       expression: TSESTree.Expression,
     ): expression is TSESTree.Identifier | TSESTree.StringLiteral {
-      const type = DEPRECATED_getConstrainedTypeAtLocation(
-        services,
-        expression,
-      );
+      const type = getConstrainedTypeAtLocation(services, expression);
 
       const isString = (t: ts.Type): boolean => {
         return isTypeFlagSet(t, ts.TypeFlags.StringLike);
