+++ packages/eslint-plugin/src/rules/no-unsafe-return.ts
@@ -7,7 +7,7 @@ import {
   AnyType,
   createRule,
   discriminateAnyType,
-  DEPRECATED_getConstrainedTypeAtLocation,
+  getConstrainedTypeAtLocation,
   getContextualType,
   getParserServices,
   getThisExpression,
@@ -68,10 +68,7 @@ export default createRule({
       }
 
       // function has an explicit return type, so ensure it's a safe return
-      const returnNodeType = DEPRECATED_getConstrainedTypeAtLocation(
-        services,
-        returnNode,
-      );
+      const returnNodeType = getConstrainedTypeAtLocation(services, returnNode);
       const functionTSNode = services.esTreeNodeToTSNodeMap.get(functionNode);
 
       // function expressions will not have their return type modified based on receiver typing
@@ -162,7 +159,7 @@ export default createRule({
           if (
             thisExpression &&
             isTypeAnyType(
-              DEPRECATED_getConstrainedTypeAtLocation(services, thisExpression),
+              getConstrainedTypeAtLocation(services, thisExpression),
             )
           ) {
             messageId = 'unsafeReturnThis';
